+++ docs/src/dokumentation/RELEASENOTES.md
@@ -15,3 +15,10 @@
 - Frontend auf Mobidam angepasst
 - Dummy-Frontend: Schnittstellen Detail-View
 - Zuordnung (Erstellung), Anzeigen und Löschen von Personen zu Schnittstellen in der UI
+
+## Version 3.0 - 16.01.2024
+### Hinzugefügt
+- Tabelle Schnittstelle mit `Create`, `Update`, `GetAll`, `Delete`
+
+### Verbessert
+- Beziehung zwischen Zuordnungen und Schnittstellen eingefügt